[FIX] Updates size and index types to comply with PEP 353

+++ cloc/parsing/extensions/_comment_data.c
@@ -3,7 +3,7 @@
 
 void initialize_comment_data(struct CommentData *comment_data,
     const char *singleline_symbol, const char *multiline_start_symbol, const char *multiline_end_symbol,
-    unsigned int singleline_length, unsigned int multiline_start_length, unsigned int multiline_end_length){
+    Py_ssize_t singleline_length, Py_ssize_t multiline_start_length, Py_ssize_t multiline_end_length){
 
     comment_data->singleline_length = singleline_length;
     comment_data->multiline_start_length = multiline_start_length;
